fix(sheet): fix numfmt storybook error
